Split InternalTradeArea to InternalTradeArea, DistrictProductionInfo, DistrictProductionInfoComparator, TradeDistrictProcessingInfo.
