+++ sources/Colonization/AngelScriptBinders/Core/InternalTradeArea/BindDistrictProductionInfoKeys.hpp
@@ -1,0 +1,6 @@
+#pragma once
+#include <Urho3D/AngelScript/Script.h>
+namespace Colonization
+{
+void BindDistrictProductionInfoKeys (Urho3D::Script *script);
+}
