+++ sources/Colonization/AngelScriptBinders/Core/InternalTradeArea/BindInternalTradeArea.hpp
@@ -3,7 +3,5 @@
 namespace Colonization
 {
 void BindInternalTradeArea (Urho3D::Script *script);
-void BindDistrictProductionKeys (Urho3D::Script *script);
-void BindTradeDistrictsProcessingInfo (Urho3D::Script *script);
 void BindInternalTradeAreaInterface (Urho3D::Script *script, Urho3D::String className);
 }
