+++ sources/Colonization/Backend/VictoryProgressUpdater.hpp
@@ -3,7 +3,7 @@
 #include <Urho3D/AngelScript/ScriptFile.h>
 
 #include <Colonization/Core/Map.hpp>
-#include <Colonization/Core/InternalTradeArea.hpp>
+#include <Colonization/Core/InternalTradeArea/InternalTradeArea.hpp>
 #include <Colonization/Core/GameConfiguration.hpp>
 
 #include <Colonization/Backend/PlayersManager.hpp>
