+++ sources/Colonization/Core/InternalTradeArea/DistrictProductionInfoComparator.cpp
@@ -1,0 +1,13 @@
+#include <Colonization/BuildConfiguration.hpp>
+#include "DistrictProductionInfoComparator.hpp"
+
+namespace Colonization
+{
+namespace DistrictProductionInfoComparators
+{
+bool HigherSellability (DistrictProductionInfo &first, DistrictProductionInfo &second)
+{
+    return (first.sellability_ > second.sellability_);
+}
+}
+}
