+++ sources/Colonization/Core/InternalTradeArea/InternalTradeArea.cpp
@@ -900,112 +900,4 @@ void InternalTradeArea::SetDistrictsHashesArrayAttribute (const Urho3D::VariantV
         }
     }
 }
-
-TradeDistrictProcessingInfo::TradeDistrictProcessingInfo (Urho3D::Context *context) : Urho3D::Object (context),
-    unusedProduction_ (),
-    unsoldTradeGoodsCost_ (0.0f),
-    soldTradeGoodsCost_ (0.0f)
-{
-
-}
-
-TradeDistrictProcessingInfo::~TradeDistrictProcessingInfo ()
-{
-
-}
-
-float TradeDistrictProcessingInfo::GetUnusedProductionOf (Urho3D::StringHash type) const
-{
-    float *unusedProduction = unusedProduction_ [type];
-    if (unusedProduction)
-    {
-        return *unusedProduction;
-    }
-    else
-    {
-        return 0.0f;
-    }
-}
-
-void TradeDistrictProcessingInfo::SetUnusedProductionOf (Urho3D::StringHash type, float points)
-{
-    unusedProduction_ [type] = points;
-}
-
-Urho3D::VariantMap TradeDistrictProcessingInfo::GetDistrictProduction (Urho3D::StringHash districtHash) const
-{
-    Urho3D::VariantMap *districtProduction = districtsProduction_ [districtHash];
-    if (districtProduction)
-    {
-        return *districtProduction;
-    }
-    else
-    {
-        return Urho3D::Variant::emptyVariantMap;
-    }
-}
-
-void TradeDistrictProcessingInfo::SetDistrictProduction (Urho3D::StringHash districtHash, Urho3D::VariantMap &districtProduction)
-{
-    districtsProduction_ [districtHash] = districtProduction;
-}
-
-Urho3D::VariantMap TradeDistrictProcessingInfo::GetDistrictBalanceAdditions (Urho3D::StringHash districtHash) const
-{
-    Urho3D::VariantMap *districtBalanceAdditions = districtsBalanceAdditions_ [districtHash];
-    if (districtBalanceAdditions)
-    {
-        return *districtBalanceAdditions;
-    }
-    else
-    {
-        return Urho3D::Variant::emptyVariantMap;
-    }
-}
-
-void TradeDistrictProcessingInfo::SetDistrictBalanceAdditions (Urho3D::StringHash districtHash, Urho3D::VariantMap &districtBalanceAdditions)
-{
-    districtsBalanceAdditions_ [districtHash] = districtBalanceAdditions;
-}
-
-float TradeDistrictProcessingInfo::GetUnsoldTradeGoodsCost () const
-{
-    return unsoldTradeGoodsCost_;
-}
-
-void TradeDistrictProcessingInfo::SetUnsoldTradeGoodsCost (float unsoldTradeGoodsCost)
-{
-    unsoldTradeGoodsCost_ = unsoldTradeGoodsCost;
-}
-
-float TradeDistrictProcessingInfo::GetSoldTradeGoodsCost () const
-{
-    return soldTradeGoodsCost_;
-}
-
-void TradeDistrictProcessingInfo::SetSoldTradeGoodsCost (float soldTradeGoodsCost)
-{
-    soldTradeGoodsCost_ = soldTradeGoodsCost;
-}
-
-bool DistrictProductionInfoComparators::HigherSellability (DistrictProductionInfo &first, DistrictProductionInfo &second)
-{
-    return (first.sellability_ > second.sellability_);
-}
-
-void DistrictProductionInfo::CalculateSellability ()
-{
-    sellability_ = quality_ / relativePrice_;
-}
-
-Urho3D::VariantMap DistrictProductionInfo::ToVariantMap ()
-{
-    Urho3D::VariantMap variantMap;
-    variantMap [DISTRICT_PRODUCTION_AMOUNT_KEY] = amount_;
-    variantMap [DISTRICT_PRODUCTION_RELATIVE_PRICE_KEY] = relativePrice_;
-    variantMap [DISTRICT_PRODUCTION_QUALITY_KEY] = quality_;
-    variantMap [DISTRICT_PRODUCTION_SELLABILITY_KEY] = sellability_;
-    variantMap [DISTRICT_PRODUCTION_SELLED_KEY] = selled_;
-    return variantMap;
-}
 }
